Revert "Ryan profile"
