+++ client/src/pages/Login.js
@@ -61,61 +61,6 @@ class SignInSide extends Component {
           {/* <img src={weights} width="150" height="150" /> */}
           <img src={weights} width="400" alt="weights placeholder" />
 
-
-        <Grid item xs={12} sm={4} md={5} component={Paper} elevation={6} square>
-          <MyPaperLogin>
-            <Typography component="h1" variant="h5">
-              Sign in
-            </Typography>
-            <MyFormLogin>
-              <TextField
-                variant="outlined"
-                margin="normal"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                value={this.state.username}
-                onChange={this.handleInputChange}
-                name="username"
-                autoComplete="email"
-                autoFocus
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                required
-                fullWidth
-                value={this.state.password}
-                onChange={this.handleInputChange}
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                autoComplete="current-password"
-              />
-              <FormControlLabel
-                control={<Checkbox value="remember" color="primary" />}
-                label="Remember me"
-              />
-              <Button
-                type="submit"
-                fullWidth
-                variant="contained"
-                color="primary"
-                onClick={this.handleFormSubmit}
-              >
-                Sign In
-              </Button>
-              {fireRedirect && <Redirect to={from || "/admin"} />}
-              <GoogleLogin
-                clientId="527119401095-flf80nen6cgthtcso628cp0tg59tmcij.apps.googleusercontent.com"
-                buttonText="Login"
-                onSuccess={responseGoogle}
-                onFailure={responseGoogle}
-                cookiePolicy={"single_host_origin"}
-              />
-
           <Grid
             item
             xs={12}
@@ -177,7 +122,6 @@ class SignInSide extends Component {
                   onFailure={responseGoogle}
                   cookiePolicy={"single_host_origin"}
                 />
-
 
                 <Grid container>
                   <Grid item xs>
